// project-fixtures/react-apollo-prj/src/__generated__/app-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
import { TypedDocumentNode } from "@graphql-typed-document-node/core";
export type AppQuery = {
    viewer: {
        repositories: {
            nodes: (({
                id: string;
            } & RepositoryItem_Repository) | null)[] | null;
        };
    };
};
export type AppQueryVariables = {
    first: number;
};
export type AppQueryDocument = TypedDocumentNode<AppQuery, AppQueryVariables>;
export type RepositoryItem_Repository = {
    name: string;
    description: string | null;
};